Automatic table generation for search
Search data with dynamically-generated MongoDB querying condition will be rendered in the table dynamically. Additional information not directly showing on the table will be linked to the button that will trigger Swal.Fire alert showing the entire information. Pagination and ordering in MongoDB-side(server-side) will be the further requirement to prevent lagging in case of numerous amounts of result from the database.